Enable syntax highlighting for all/Update dependencies
This commit is the seemingly odd combo of simply updating dependencies,
and also refactoring to enable syntax highlighting in code blocks for
all renders, not just for folks who have github-linguist installed.

They go together because the latest `html-pipeline` now uses `rouge`
rather than `github-linguist`.  `rouge` is a pure ruby library without
the install vagaries of `github-linguist` in its syntax highlighting
filter, so we remove the code supporting optional syntax highlighting
and move to simply supporting it for all renders.